Special spawn script required to start the spins
The original spawn file was created for the runs, where we’re starting
from a history file.  Using that script to start a spin ends up
removing all of the years, which is very bad.  The lazy hack is to just
create a new script that doesn’t have the updating of the start years
etc.