Stop submission loop if the site fails from the get-go
in instances where the site won’t even begin to run because of a
problem (often land/sea mask), we need to make sure it never tries
again.  If the model can’t even begin to run, the only thing that
should be in it’s histo folder is the .xml file that gets generated
before even the met gets read in.